respaldo/requirements.txt
aiofiles==24.1.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.10.4
bidict==0.23.1
billiard==4.2.1
blinker==1.9.0
celery==5.5.3
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
dnspython==2.7.0
dotenv==0.9.9
eventlet==0.40.1
execnet==2.1.1
Flask==3.1.0
flask-cors==6.0.0
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.58.4
gevent==24.2.1
gevent-websocket==0.10.1
git-filter-repo==2.47.0
greenlet==3.1.1
gunicorn==22.0.0
h11==0.16.0
h2==4.2.0
holidays==0.75
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
kiwisolver==1.4.8
kombu==5.5.4
MarkupSafe==3.0.2
matplotlib==3.10.3
MouseInfo==0.1.3
numpy==2.3.1
openai==1.90.0
packaging==25.0
pandas==2.2.2
pillow==10.4.0
playwright==1.53.0
playwright-stealth==2.0.0
pluggy==1.6.0
priority==2.0.0
prompt_toolkit==3.0.51
psutil==5.9.8
psycopg2-binary==2.9.9
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydot==4.0.0
pyee==13.0.0
PyGetWindow==0.0.9
Pygments==2.19.1
PyMsgBox==1.0.9
pyparsing==3.2.3
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytest==8.4.0
pytest-base-url==2.1.0
pytest-flask==1.3.0
pytest-mock==3.14.0
pytest-playwright==0.5.0
pytest-xdist==3.8.0
pytest-xprocess==0.23.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.12.2
python-json-logger==3.3.0
python-slugify==8.0.4
python-socketio==5.13.0
pytweening==1.2.0
pytz==2025.2
pywin32==310
Quart==0.20.0
redis==6.2.0
requests==2.32.3
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.31
sqlalchemy_schemadisplay==2.0
text-unidecode==1.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wsproto==1.2.0
WTForms==3.2.1
zope.event==5.0
zope.interface==6.4.post2
